Review test scripts format
